Actually commit the code for this hack.
jmap doesn't work too well if the corefile being analyzed is greater
than 2G in size, as it uses the "small" file open. This hack basically
replaces the I/O routines in the standard library so that they take treat
offsets as unsigned. This is not globally correct in any shape or form,
but the usage pattern of jmap is such that it does actually work.